requirements.txt

Flask
redis
pyyaml